module/xep0313: mam manager
